feature: hooks parameters
